reorganization and cleanup for the different validators
