--- SchemaValidator/xml/good.hml.1.0.1.xml
@@ -1,122 +1,0 @@
-<?xml version="1.0" ?>
-<hml xmlns="http://schemas.nmdp.org/spec/hml/1.0.1"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://schemas.nmdp.org/spec/hml/1.0.1  http://schemas.nmdp.org/spec/hml/1.0.1/hml-1.0.1.xsd"
-    version="1.0.1" >
-    <hmlid root="2.34.48.32" extension="HML.3245662"/>
-    <reporting-center reporting-center-id="567"/>
-    <sample id="4555-6677-8">
-        <typing gene-family="HLA" date="2015-01-13">
-            <allele-assignment date="2015-07-28" allele-db="IMGT/HLA" allele-version="3.17.0">
-                <haploid locus="HLA-A" method="DNA" type="02:20:01"/>
-                <glstring>
-                    HLA-A*02:20:01
-                </glstring>
-            </allele-assignment>
-            <typing-method>
-               <sbt-ngs locus="HLA-A"
-                    test-id="HLA-A.Test.1234"
-                    test-id-source="AcmeGenLabs">
-                    <raw-reads uri="rawreads/read1.fastq.gz"
-                    availability="public"
-                    format="fastq"
-                    paired="1"
-                    pooled="1"
-                    adapter-trimmed="1"
-                    quality-trimmed="0"/>
-                </sbt-ngs>
-            </typing-method>
-            <consensus-sequence date="2015-01-13">
-
-       
-                <reference-database availability="public" curated="true">
-                    <reference-sequence
-                    name="HLA-A reference"
-                    id="Ref111"
-                    start="945000"
-                    end="946000"
-                    accession="GL000123.4"
-                    uri="http://AcmeGenReference/RefDB/GL000123.4"/>
-                </reference-database>
-
-          
-                <consensus-sequence-block reference-sequence-id="Ref111"
-                    start="945532"
-                    end="945832"
-                    strand="+"
-                    phase-set="1"
-                    expected-copy-number="1"
-                    continuity="true"
-                    description="HLA-A Consensus Sequence 4.5.67">
-
-        
-                    <sequence>
-                        CCCAGTTCTCACTCCCATTGGGTGTCGGGTTTCCAGAGAAGCCAATCAGTGTCGTCGCGGTCGCTGTTCTAAAGCCCGCACGCACCCACCGGGACTCAGATTCTCCCCAGACGCCGAGGATGGCCGTCATGGCGCCCCGAACCCTCCTCCTGCTACTCTCGGGGGCCCTGGCCCTGACCCAGACCTGGGCGGGTGAGTGCGGGGTCGGGAGGGAAACCGCCTCTGCGGGGAGAAGCAAGGGGCCCTCCTGGCGGGGGCGCAGGACCGGGGGAGCCGCGCCGGGACGAGGGTCGGGCAGGT
-                    </sequence>
-
-
-                    <variant id="0"
-                        reference-bases="GTCATG"
-                        alternate-bases="ACTCCC"
-                        start="945542"
-                        end="945548"
-                        filter="pass"
-                        quality-score="95">
-
-
-                        <variant-effect term="missense_variant"/>
-                    </variant>
-                </consensus-sequence-block>
-            </consensus-sequence>
-        </typing>
-    </sample>
-
-    <sample id="4555-6677-9">
-        <typing gene-family="HLA" date="2015-01-13">
-            <allele-assignment date="2015-07-28" allele-db="IMGT/HLA" allele-version="3.17.0">
-                <haploid locus="HLA-A" method="DNA" type="02:20:01"/>
-                <glstring>
-                    HLA-A*02:01:01:01
-                </glstring>
-            </allele-assignment>
-            <typing-method>
-                <sbt-ngs locus="HLA-A"
-                    test-id="HLA-A.Test.1234"
-                    test-id-source="AcmeGenLabs">
-                    <raw-reads uri="rawreads/read2.fastq.gz"
-                    availability="public"
-                    format="fastq"
-                    paired="1"
-                    pooled="1"
-                    adapter-trimmed="1"
-                    quality-trimmed="0"/>
-                </sbt-ngs>
-            </typing-method>
-            <consensus-sequence date="2015-01-13">
-                <reference-database availability="public" curated="true">
-                    <reference-sequence
-                    name="HLA-A reference"
-                    id="Ref112"
-                    start="945000"
-                    end="946000"
-                    accession="GL000123.4"
-                    uri="http://AcmeGenReference/RefDB/GL000123.4"/>
-                </reference-database>
-                <consensus-sequence-block
-                    reference-sequence-id="Ref112"
-                    start="945532"
-                    end="945832"
-                    strand="+"
-                    phase-set="1"
-                    expected-copy-number="1"
-                    continuity="true"
-                    description="HLA-A Consensus Sequence 4.5.89">
-                    <sequence>
-                        CCCAGTTCTCGTCATGATTGGGTGTCGGGTTTCCAGAGAAGCCAATCAGTGTCGTCGCGGTCGCTGTTCTAAAGCCCGCACGCACCCACCGGGACTCAGATTCTCCCCAGACGCCGAGGATGGCCGTCATGGCGCCCCGAACCCTCCTCCTGCTACTCTCGGGGGCCCTGGCCCTGACCCAGACCTGGGCGGGTGAGTGCGGGGTCGGGAGGGAAACCGCCTCTGCGGGGAGAAGCAAGGGGCCCTCCTGGCGGGGGCGCAGGACCGGGGGAGCCGCGCCGGGACGAGGGTCGGGCAGGT
-                    </sequence>
-                </consensus-sequence-block>
-            </consensus-sequence>
-        </typing>
-    </sample>
-
-</hml>
